Adding UUM logging to medications
