+++ src/applications/mhv-medications/containers/Prescriptions.jsx
@@ -13,6 +13,8 @@ import PropTypes from 'prop-types';
 import {
   usePrintTitle,
   updatePageTitle,
+  logUniqueUserMetricsEvents,
+  EVENT_REGISTRY,
 } from '@department-of-veterans-affairs/mhv/exports';
 import { isAuthenticatedWithSSOe } from '~/platform/user/authentication/selectors';
 import MedicationsList from '../components/MedicationsList/MedicationsList';
@@ -317,6 +319,26 @@ const Prescriptions = () => {
       updatePageTitle('Medications | Veterans Affairs');
     },
     [currentPage],
+  );
+
+  // Log when prescriptions are successfully displayed to the user
+  useEffect(
+    () => {
+      if (
+        !isPrescriptionsLoading &&
+        !isPrescriptionsFetching &&
+        !prescriptionsApiError &&
+        prescriptionsData
+      ) {
+        logUniqueUserMetricsEvents(EVENT_REGISTRY.PRESCRIPTIONS_ACCESSED);
+      }
+    },
+    [
+      isPrescriptionsLoading,
+      isPrescriptionsFetching,
+      prescriptionsApiError,
+      prescriptionsData,
+    ],
   );
 
   // Update loading state based on RTK Query states
